+++ cpp/src/rolling/jit/code/code.h
@@ -20,7 +20,6 @@
 #pragma once
 
 namespace cudf {
-namespace experimental {
 namespace rolling {
 namespace jit {
 namespace code {
@@ -37,5 +36,4 @@ extern const char* grouped_window_wrapper;
 }  // namespace code
 }  // namespace jit
 }  // namespace rolling
-}  // namespace experimental
 }  // namespace cudf