+++ cpp/tests/groupby/group_count_test.cpp
@@ -36,7 +36,7 @@ TYPED_TEST(groupby_count_test, basic)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::COUNT_VALID>;
+    using R = cudf::detail::target_type_t<V, aggregation::COUNT_VALID>;
 
     fixed_width_column_wrapper<K> keys { 1, 2, 3, 1, 2, 2, 1, 3, 3, 2};
     fixed_width_column_wrapper<V> vals { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
@@ -44,13 +44,13 @@ TYPED_TEST(groupby_count_test, basic)
     fixed_width_column_wrapper<K> expect_keys { 1, 2, 3 };
     fixed_width_column_wrapper<R> expect_vals { 3, 4, 3 };
 
-    auto agg = cudf::experimental::make_count_aggregation();
+    auto agg = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg1 = cudf::experimental::make_count_aggregation();
+    auto agg1 = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg1), force_use_sort_impl::YES);
 
-    auto agg2 = cudf::experimental::make_count_aggregation(null_policy::INCLUDE);
+    auto agg2 = cudf::make_count_aggregation(null_policy::INCLUDE);
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg2));
 }
 
@@ -58,7 +58,7 @@ TYPED_TEST(groupby_count_test, empty_cols)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::COUNT_VALID>;
+    using R = cudf::detail::target_type_t<V, aggregation::COUNT_VALID>;
 
     fixed_width_column_wrapper<K> keys        { };
     fixed_width_column_wrapper<V> vals        { };
@@ -66,10 +66,10 @@ TYPED_TEST(groupby_count_test, empty_cols)
     fixed_width_column_wrapper<K> expect_keys { };
     fixed_width_column_wrapper<R> expect_vals { };
 
-    auto agg = cudf::experimental::make_count_aggregation();
+    auto agg = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg1 = cudf::experimental::make_count_aggregation();
+    auto agg1 = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg1), force_use_sort_impl::YES);
 }
 
@@ -77,7 +77,7 @@ TYPED_TEST(groupby_count_test, zero_valid_keys)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::COUNT_VALID>;
+    using R = cudf::detail::target_type_t<V, aggregation::COUNT_VALID>;
 
     fixed_width_column_wrapper<K> keys( { 1, 2, 3}, all_null() );
     fixed_width_column_wrapper<V> vals  { 3, 4, 5};
@@ -85,13 +85,13 @@ TYPED_TEST(groupby_count_test, zero_valid_keys)
     fixed_width_column_wrapper<K> expect_keys { };
     fixed_width_column_wrapper<R> expect_vals { };
 
-    auto agg = cudf::experimental::make_count_aggregation();
+    auto agg = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg1 = cudf::experimental::make_count_aggregation();
+    auto agg1 = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg1), force_use_sort_impl::YES);
 
-    auto agg2 = cudf::experimental::make_count_aggregation(null_policy::INCLUDE);
+    auto agg2 = cudf::make_count_aggregation(null_policy::INCLUDE);
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg2));
 }
 
@@ -99,7 +99,7 @@ TYPED_TEST(groupby_count_test, zero_valid_values)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::COUNT_VALID>;
+    using R = cudf::detail::target_type_t<V, aggregation::COUNT_VALID>;
 
     fixed_width_column_wrapper<K> keys   { 1, 1, 1};
     fixed_width_column_wrapper<V> vals ( { 3, 4, 5}, all_null() );
@@ -107,14 +107,14 @@ TYPED_TEST(groupby_count_test, zero_valid_values)
     fixed_width_column_wrapper<K> expect_keys { 1 };
     fixed_width_column_wrapper<R> expect_vals { 0 };
 
-    auto agg = cudf::experimental::make_count_aggregation();
+    auto agg = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg1 = cudf::experimental::make_count_aggregation();
+    auto agg1 = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg1), force_use_sort_impl::YES);
 
     fixed_width_column_wrapper<R> expect_vals2 { 3 };
-    auto agg2 = cudf::experimental::make_count_aggregation(null_policy::INCLUDE);
+    auto agg2 = cudf::make_count_aggregation(null_policy::INCLUDE);
     test_single_agg(keys, vals, expect_keys, expect_vals2, std::move(agg2));
 }
 
@@ -122,7 +122,7 @@ TYPED_TEST(groupby_count_test, null_keys_and_values)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::COUNT_VALID>;
+    using R = cudf::detail::target_type_t<V, aggregation::COUNT_VALID>;
 
     fixed_width_column_wrapper<K> keys({ 1, 2, 3, 1, 2, 2, 1, 3, 3, 2, 4},
                                        { 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1});
@@ -134,14 +134,14 @@ TYPED_TEST(groupby_count_test, null_keys_and_values)
                                           //  { 3, 6,     1, 4, 9,   2, 8,    -}
     fixed_width_column_wrapper<R> expect_vals { 2,        3,         2,       0};
 
-    auto agg = cudf::experimental::make_count_aggregation();
+    auto agg = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg1 = cudf::experimental::make_count_aggregation();
+    auto agg1 = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg1), force_use_sort_impl::YES);
 
     fixed_width_column_wrapper<R> expect_vals2{ 3,        4,         2,       1};
-    auto agg2 = cudf::experimental::make_count_aggregation(null_policy::INCLUDE);
+    auto agg2 = cudf::make_count_aggregation(null_policy::INCLUDE);
     test_single_agg(keys, vals, expect_keys, expect_vals2, std::move(agg2));
     
 }
@@ -153,7 +153,7 @@ TEST_F(groupby_count_string_test, basic)
 {
     using K = int32_t;
     using V = cudf::string_view;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::COUNT_VALID>;
+    using R = cudf::detail::target_type_t<V, aggregation::COUNT_VALID>;
 
     fixed_width_column_wrapper<K> keys        {   1,   3,   3,   5,   5,   0};
     strings_column_wrapper        vals        { "1", "1", "1", "1", "1", "1"};
@@ -161,10 +161,10 @@ TEST_F(groupby_count_string_test, basic)
     fixed_width_column_wrapper<K> expect_keys   {   0,   1,   3,   5};
     fixed_width_column_wrapper<R> expect_vals   {   1,   1,   2,   2};
 
-    auto agg = cudf::experimental::make_count_aggregation();
+    auto agg = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg1 = cudf::experimental::make_count_aggregation();
+    auto agg1 = cudf::make_count_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg1), force_use_sort_impl::YES);
 }
 // clang-format on