+++ CHANGELOG.md
@@ -201,6 +201,8 @@
 - PR #5208 Port search and join benchmark to libcudf++
 - PR #5214 Move docs build script into repository
 - PR #5219 Add per context cache for JIT kernels
+- PR #5233 Remove experimental namespace used during libcudf++ refactor
+- PR #5251 Fix more mispellings in cpp comments and strings
 
 ## Bug Fixes
 
@@ -326,6 +328,7 @@
 - PR #5212 Fix memory leak in `dlpack.pyx:from_dlpack()`
 - PR #5224 Add new headers from 5198 to libcudf/meta.yaml
 - PR #5228 Fix datetime64 scalar dtype handling for unsupported time units
+- PR #5256 ORC reader: fix loading individual timestamp columns
 
 
 # cuDF 0.13.0 (31 Mar 2020)